fullstack-orthanc-dicom/simrs-order-ui/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
httpx==0.27.2
pydantic==2.9.2
python-dotenv==1.0.1
orjson==3.10.7
jinja2==3.1.4
rich==13.7.1
psycopg2-binary==2.9.9
requests==2.32.3